Phase 6a closes the two outstanding PRD CLI line items (`update`,
`config`), adds opt-in endpoint authentication, and drains 13 backlog
items from Phase 5 live testing. Doctor 12 → 14 checks. ParamsB
migrated from int to float64 for sub-1B model precision. New
PreferredIDs: qwen3-0.6b, qwen3-1.7b.

Live smoke caught and fixed a Phase-5-deferred bug: portConflictsCheck
was using net.Listen as the in-use probe, which is unreliable on macOS
due to Go's default SO_REUSEADDR. Switched to net.DialTimeout.

+++ internal/models/metadata.go
@@ -15,8 +15,8 @@ type Metadata struct {
 
 	// Phase 2.5 additions; omitempty so legacy files decode cleanly
 	// (legacy entries lack these — selector will not run on them).
-	ParamsB int  `json:"params_b,omitempty"`
-	Arch    Arch `json:"arch,omitempty"`
+	ParamsB float64 `json:"params_b,omitempty"`
+	Arch    Arch    `json:"arch,omitempty"`
 
 	// Phase 3 addition. Updated by `serve` (foreground or detached)
 	// immediately before launching llama-server.
